+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add and deploy Vitepress documentation
 - Add support for [Contributors list](./README.md#contributors-) using [all-contributors](https://github.com/all-contributors/all-contributors) specification
 - Add onboarding of community member after setup.
+- Add Privacy Notice page /privacy
+- Add Community Terms of Service page /terms
 
 ### Changed